More cleanup

+++ src/main/java/com/naterbobber/darkerdepths/client/particle/ColoredAshParticle.java
@@ -22,9 +22,9 @@ public class ColoredAshParticle extends BaseAshSmokeParticle {
 
     protected static ColoredAshParticle coloredAshParticle(ClientLevel level, double x, double y, double z, float quadSizeMultiplier, SpriteSet sprites, float rCol, float gCol, float bCol, BrightnessBehavior brightnessBehavior) {
         RandomSource randomsource = level.random;
-        double d0 = (double)randomsource.nextFloat() * -1.9 * (double)randomsource.nextFloat() * 0.1;
-        double d1 = (double)randomsource.nextFloat() * (double)-0.5F * (double)randomsource.nextFloat() * 0.1 * (double)5.0F;
-        double d2 = (double)randomsource.nextFloat() * -1.9 * (double)randomsource.nextFloat() * 0.1;
+        double d0 = randomsource.nextDouble() * -1.9 * randomsource.nextDouble() * 0.1;
+        double d1 = randomsource.nextDouble() * -0.5 * randomsource.nextDouble() * 0.1 * 5.0;
+        double d2 = randomsource.nextDouble() * -1.9 * randomsource.nextDouble() * 0.1;
         return new ColoredAshParticle(level, x, y, z, d0, d1, d2, quadSizeMultiplier * 1.5F, sprites, rCol, gCol, bCol, brightnessBehavior);
     }
